- Refactored server startup
