+++ README.md
@@ -1,4 +1,3 @@
 ## Lifehub
 Front End: React⚛️
-
-Back End: Currently Firebase, will refactor to use TypeORM, GQL and Apollo
+Back End: TypeORM, GQL and Apollo
