+++ server/package.json
@@ -2,8 +2,10 @@
    "name": "lifehub",
    "version": "0.0.1",
    "dependencies": {
+      "@types/bcryptjs": "^2.4.2",
       "apollo-server": "^2.11.0",
       "bcrypt": "^4.0.1",
+      "bcryptjs": "^2.4.3",
       "config": "^3.3.1",
       "dotenv": "^8.2.0",
       "graphql": "^15.0.0",
